BUG:  Fixed minor differences in testing.
Some of the tests were failed because there were minor differences
between baselines and currently generated outputs. This happened
because now BCD uses ITK4 instead of ITK3, which causes small
differences in the intensity of the output images.
Therefore, I replaced the previous baselines with the new ones
which differ very little. Now all the tests relating to the following
baselines are passed.