BUG: Improper formatting for conversion strings.
http://www.na-mic.org/Bug/view.php?id=1464

I believe I found the issue here - it was a very odd failure that defied direct debugging in gdb or visual studio because this code died deep in the PARSE_ARGS macro. But after some looking around I saw that one of the command line arguments was oddly formatted -- in fact, you can see if in the earlier notes of this bug report:

--removeIntensityOutliers 0.0F

It turns out that the F is included in the default in the BRAINSFit.xml file but if I remove it (see diff below) then test runs.