Adding new programs to compute the landmarks statistics
Two new programs called "GenerateAverageLmkFile.cxx" and
"LmkStatistics.cxx" are put in a new directory called
"BRAINSConstellationDetector/landmarkStatistics/". The first program
computes the average between some landmark files, and the second program
is used for comparing several landmark files (fcsv files). This program
extract the average, variance and standard deviation of the errors
between several pair of fcsv files. The bigenning of each program is
fully commented which describes the use of them.