BUG: using 'end' as a function name made XMLParser crash
Apparently, 'end' is a name already in use at the point where the Expat parser
is given it's start and end handlers.  This other 'end' gets passed in instead
of the real function we want as an end element handler, and consequently at the
end of the first XML element, the Expat parser tried to call a non-existent function
causing the crash.  This seemed to be a Linux-only problem.