Fix BRAINSConstellationModeler to use correct acpc-aligned transform
Before BRAINSConstellationModeler was using reflective correlation
method to transform points to the acpc-aligned space, but based on the
information printed on screen, points did not transform correctly. Now I
estimated the transform using the estimation of the plane passing
through the RP, AC and PC points. I printed out the results and now the
transformation works properly.