COMP: Fix the failing landmarksWeightsCompareTest
Previously a good estimation for MSP plane happens when the Reflective correlation
value between two parts of the brains is more than 0.64.

I changed this threshold to 0.64. The new treshold is chosen by some
statistical calculation on 23 successfully passed data.

Now, using the new threshold, the results are exactly the same, and the
test again passes successfully.