ENH: Adding histogram equalization option to ImageCalculator
ImageCalculator did not return meaningful results when the intensity
values of the input images has a DC difference. Therefore, a new option is
added to ImageCalculator called "ifhisteq".
If the accumulator buffer is not empty, then every subsequent image is
histogram equalized to the current accumulator buffer.
Usage of the new option:
-ifhisteq <NumberOfMatchPoints>