COMP:  Install ITKv4 and VTK to reduce include paths.
When pointing ITK_DIR or VTK_DIR to a build tree, the include paths
are very long.  By installing these packages, the include paths become
significantly shorter.

This facilitates several development tools that need to search for includes
to work faster (including compilers themselves, but clangcomplete and cppcheck
in particular).