COMP: Avoid continuous rebuild of ExternalPackages
New option added to that defaults to off for FORCE_EXTERNAL_BUILDS.
This needs to be turned ON to allow updating of external packages.

Needed to split the ExternalProject command and argument into separate
varables for their effect to work.  Otherwise they were having no
noticiable effect on the build system.