COMP: Avoid slow recompilation of external projects.
This patch adds a cmake option to dictate weather the
external packages are updated (built, and reinstalled)
with every run of make, or weather it simply assumes
that once an external package is built that it remains
static and falls through quickly.

The previous implementation was not working as desired
and DCMTK, Boost, (and all the others) were being updated
configured and built every time.  This process was taking
several minutes with no benefit.

If one does update a tag, then changing the FORCE_EXTERNAL_BUILDS
to ON for 1 build cycle, and then back off will update all
the necessary dependancies.