COMP:  Compiler warning in BRAINSCut indicated bug.
+::GetOneConnectedRegion( LabelImagePointerType& image, const unsigned char labelValue)

labelValue was not used, but then in that function the resultMask = ExtractLabel( multipleLabelVolume, 1 );
was hard coded to 1.  I think this was an incomplete refactoring where the labelValue is the desired
value to extract.

I also removed default paramters because they make the code difficult to read, and maintaining
backwards compatibility is not necessary.