ENH: Add capabilities to write out csv log file in a report
It is nice to be able to write out the metric value in a CSV file for
knowing the quality of registration results for various metrics.

This feature fits nicely with BRAINSFit tools.

The new command line option is
--logFileReport

A file to write out final information report in CSV file:
FixedImageName,FixedMaskName,MovingImageName,MovingMaskName,MetricName,MetricValue