ENH: Remove DeformationField parameter from GenericTransformImage
See https://www.icts.uiowa.edu/jira/browse/PREDICTIMG-1667

The GenericTransformImage template function isn't used that many
places. In most it's clear that either a DeformationField or
a Transform is passed in. A few places in BRAINSCut it looks like
both are passed in, but in practice only one or the other is
actually defined.

The way GenericTransformImage was actually written, if both a
transform and a displacement field are passed in, the transform
is ignored.  So I felt justified in enforcing that at the
point of call.  If needed, a DisplaceFieldTransform is created
and the Displacement Image is passed in.

The biggest 'hidden' change this brings about is that the
DisplacementField type is defined as
itk::Image<itk::Vector<float,3>,3> but the transform parameter
type is double, so I changed type DisplacementField type so it
was compatible with the Transform type.