BUG: Landmark compare was unstable.
The LandmarkCompare program was comparing strings instead of landmarks.
That was very unstable and had no way to test for tolerance.  It also
used much custom code rather than the landmark readers/writers.

This was completely re-written to to reduce code, provide more complete
code coverage, and to add a tolerance level to allow small floating point
deviations in the landmark locations.