COMP: Replace linear image list with map of lists by type.
In order to re-do the algorithm to use covariance between
image types, first we need to change the datastructure for
input images & derivatives from a simple vector to a collection
of vectors organized by type.

This is in preparation for re-writing ComputeDistributions.h such
that the covariance is computed between types of images instead of
between individual images. This will, it is hoped, remove the
error condition when for example 2 T1 images are given as input
that have very small differences.