ENH: force base landmark points to BCD by commandline
BCD has four flags that help user to manually specify base landmark
points (RP,AC,PC,VN4) from original image on the command line.

These flags are called:
--forceRPPoint
--forceACPoint
--forcePCPoint
--forceVN4Point

Although the above flags exist at the current version of BCD, there is
no implementation for them, and using them has no effect on the final
results. In this patch, BCD is changed in a way that can consider these
flags in the landmark detection phase.

Then 8 new tests are added to BCD, which test the implementation of these
new features.

All the BCD tests are passed after the changes of this patch.