ENH: Output precision type is added to BRAINSTransformConvert
A new flags are added to BRAINSTransformConvert tool:
--outputPrecisionType

This flag can be set as "float" or "double", and it can work well with
other options like outputTransformType.

This patch is tested and works properly based on the last version of ITK
that includes updated transform file reader/writer filters.
The updated filters can convert their input transforms to the requested
output precision type automatically.

Testsuite is changed to add float precision tests. Number of tests are
increased from 63 to 125 and all of them are passed successfully.
"100% tests passed, 0 tests failed out of 125"