WIP: input/output precision type is added to BRAINSTransformConvert
Two new flags are added to BRAINSTransformConvert tool:
--inputPrecisionType
--outputPrecisionType

These flags can be set as float or double, and they can work well with
other options like outputTransformType.

This patch is tested and works properly now. However, it is not
finalized yet. First, ITK transform file reader/ writer should be fixed
in a way that they can convert their input transforms to the requested
output precision type automatically. Then, this program can be simplified
a lot based on the new version of ITK.