BUG: create usefull debug 2D branded image for BCD
When we use "writeBranded2DImage" flag in BCD, we want to have a 2D
debug image at the output that shows the midbrain base landmark centers
and the search area around them by a circle.

The current created 2D branded image just includes a fully colorfull
circle for the search area, but it does not show the fiducial center
point.

The current code is changed in a way so we can see usefull intended
information at the output 2D branded image.

Also, the related test baseline file is replaced based on new changes.