ENH: Update BRAINSFit registration framework to ITKv4
BRAINSFit is tested and works successfully based on the ITKv4 registration
framework.

The updated BRAINSFit also uses the benefits of newly added Versor and
LBFGSB optimizers.

The new version can handle a multi stages registration process including
linear and non-linear (BSpline, SyN) transforms. A 3 level registration
process is run on each stage.

This version collapses all linear transforms to only one before
writing the final output transform to the disk. Just Notice that the
current version writes the collapsed "ScaleVersor3D" and "ScaleSkewVersor3D"
transforms as an affine transform.

We can enhance the current version by removing the collapsing step.
This can be done if we directly initialize the internal transform of the
registration filter rather than passing the initial transform as an
MovingInitialTransform to the registration framework.

This version should be built based on the ITK tag of the following gerrit
patch:
http://review.source.kitware.com/#/c/14682/