BUG: Read gradient directions if they exists
The logic would avoid reading gradient directions
if the DiffusionDirection was "NONE", but some newer
dicom files do not populate the necessary feild as "DIFFUSION",
and the result is that all gradients ended up being interpreted
as B0's.

This re-organizes to attempt to get gradient directions if
they are available, and prefering the new Supplement49 tags
if they exists.