ENH:  Add BRAINSLabelStats
Vincent Magnotta wrote this little program to make measurements for various
projects. This has many uses for a stand-alone program that will take output
these measurements and label the output based on a file defining the label
names. This implements the ability to read fsl xml files and the ants label
definitions.

A couple of additional features are you can add an arbitrary number of columns
and values to the CSV based results. This can be used to define the subject and
scan ID or add a name for the type of measurements (e.g. FA, MD, etc). I have
also added calculation of the median and allow the user to determine how to
compute the median values based on the histogram.

It would be trivial to add a CSV format.