ENH: Remove collapsing step in BRAINSFitv4
At each stage of the previous version of BRAINSFitv4, the initial
transform was passed to the registration framework as a
"MovingInitalTransform". Then at the end of each stage, the initial
transform was collapsed to the registration result.

In the current patch, we aim to optimize and update the parameters of
the initial transform directly.
For this method, the output transform is set as a pointer to the internal
transform of the registration filter. Then before registration
starts, parameters of the internal transform are updated based on the
input initial transform, and during the registration process, optimizer
works directly on this initialized internal transform.
It is an analogy to the method that was performed on ITKv3 registration
framework.

Finally, after registration, the result transform will replace the current
initial transform, and itself will be considered for the initializtion of
the next stage.

The advantage of the above method is that it can increase the
registration precision by removing the collapsing step and avoiding the
errors resulted from composing different types of transforms together.