BUG: Local .h was overriding dynamically generated file
The BRAINSCommonLib.h must be used from the BRAINSCommonLib.h.in configured version
that exists in the binary tree.  A local copy of the file was put here by accident.

The result was the that the CMake configured options were not properly
propogated to the building of the object files.