ENH: Add ReferenceAtlas and MD5 signatures
The ReferenceAtlas was previously stored in SVN, but
that solution was very combersome and expensive (from a
disk usage and network burden cost) to maintain for
nightly builds.

By using MIDAS, all the object files can be cached
on a local mount point by "export ExternalData_OBJECT_STORES=<>"
environmental variables.