COMP: Move ITK find_package calls into each sub-project.
This commit simplifies how ITK components are associated with each
sub-project. Instead of maintaing a complex logic in 'BRAINSTools.cmake",
the list of component for each sub-project is generated using the
"WhatModulesITK.py" script.

To automatically prepend the find_package call to every sub-project and
reduce the amount of manual editing required, assuming the source of
ITKv4 are in /path/to/src/ITKv4, the following one-liner
has been used:

cd /path/to/src/BRAINSTools

for proj in `ls -d */CMakeLists.txt`; do \
  d=$(dirname $proj);
  echo Updating [$proj];
  components=$(../ITKv4/Utilities/Maintenance/WhatModulesITK.py \
    ../ITKv4 \
    $(find $d -type f));
  printf "#\n# ITK\n#\n${components}\n\n"|cat - $proj > /tmp/out && mv /tmp/out $proj;
done