COMP: Fixes tests by explicitly specifying list of ITK IO components
Since ITK IO factory are not explicitly instantiated in the code, this
commit explicitly set the list "ITK_IO_MODULES_USED" for each
sub-project.

In case the sub-project was already specifying some ITKIO* component in
his regular list, care has been taken to make sure they are not duplicated.

Remarks:

(1) The list has been copied from Slicer own requirement
and probably not all sub-projects require the complete list. For sake
of correctness, it could make sense to revisit each list and trim
the ITKIO components that are not relevant.

(2) Ideally, the documentation of each module should specify which file
format are supported. An idea would be to generate the documentation
based on the list of ITKIO specified in each CMakeLists.txt.

(3) Looking at the "diff", you may notice that the
variable "${ITK_IO_MODULES_USED}" is not systematically appended to
the list of components, this is explained because previous commit 6e8b92e7
was already specifying it due to copy/paste "mistake".