BUG: The master atlas should be specified from config file
There was an incorrect assertion that required coupling of the template to the
previous experiements CACHE directory. This was incorrect.  We explicitly want
each stage to be separage and only dependant on the RESULTS directories.

It is anticipated that the CACHE directories will be deleted soon after
the experiements are run.