STYLE: Remove global code from __init__.py
It is bad to have globals, and to rely on them being loaded with pythons
default module loading capabilities is even more confusing.  Simplyfy by puting
the variables in the smallest possible unit where they are need so that they are
scoped only to the code location where they are needed.

This patch also adds a feature to the MakeAtlasNode so that only subsets of the data grabber
are used.  This allows for smaller more granular data grabber nodes to be created.