COMP: Updated ITK tag to BRAINSiaITK
I updated the ITK tag to use the current version of BRAINSia/ITK
in a branch called NewParallelizedMattesMetricv4.

New performance improvements to the
MattesMutualInformationImageToImageMetricv4 are included in the above
branch.

We can revert the BRAINSia/ITK repository to the main repository as soon as
the following gerrit patch is merged to ITK:
http://review.source.kitware.com/#/c/16839