ENH: perform the local search of BCD in Freq domain
The "FindCandidatePoints" method is reimplemented to do the local search
in freqency domain.
First, based on the search bounding box, an ROI image is created from
the input MSP volume. Also, the landmark template intensities are
arranged inside an moving template image.
Then, for both fixed and moving images, corresponding maskes are
created, and the "MaskedFFTNormalizedCorrelationImageFilter" is used to
the final optimum match.

BCD is run faster now:
The average of previous run time: 3m40
The average of current run time: 1m40

That shows more than 2x speed increase.