BUG: Eliminate unnecessary handling of version
Started to refactor and elminiate the unnecessary complex
handling of versions.  Previously the code required users
of BCD to ensure that the versions matched before model generation
and during all subsequent calls to the various landmark programs.

This hard codes the consistent versions in a single location so
that if future changes are necessary, only one string needs to be changed
to indicate a consistent set of files.