ENH: Avereged warped intra subjects
A function is added that warps all of the intra subject images
(m_IntraSubjectOrigImageList) to the "key subject image" using
the intra subject registration transforms (m_IntraSubjectTransforms).
Then, it averages all T1s together and T2s together to provide
the T1 and T2 fixed images for the atlas to subject multi-modal registration.