WIP: subject intersection mask
branch name: BABC_IntersectionMask

This branch is an attempt to improve BRAINSABC
by incorporating intersection mask inter and intra-scans.
This feature would be helpful if a proper implementation can
happen. While I was testing the code, I realized that this
may take a several weeks of work when we needed this soon.
I kept the code for later development.

- Regina EY Kim