ENH: Improve T1 only segmentations
For T1 only images, sometimes the
globus does not have enough contrast
and it is overwhelmed by other structures
and ends up being squeezed out.  This
results in too few samples for the KNN
estimates.

In this version, if a label as too few
samples, then that prior is given increased
priority until sufficient samples are available.