BUG: MultiModal registration bug in BRAINSABC and BRAINSFit
Actually this patch includes fixing two bugs:
First, BRAINSABC was returning a failure when "restoreState" was not
defined in the command line.
The second bug was returning failure when we do not define neither
initial atlas to subject transform nor restoreState transform in
BRAINSABC, and this failure was related to a bug in BRAINSFit.
Both bugs are fixed and tested.