ENH: Add a python script to create inputLmkFile for BTalairach
This script creates a landmarks fcsv file (that includes AC, PC, SLA and
    IRP points) to be passed to the input of BRAINSTalairach.

This scripts uses T1/T2 volume and the label map image to find
coordinates of SLA and IRP points. These two points are used by
BRAINSTalairach to find a box around "cerebrum" part of the human brain.
Also, this script uses the output landmarks file of BCD to find the
coordinates of AC and PC points.

USAGE:

BRAINSTalairachext.py \
 --inputVolume <input T1/T2 Volume> \
 --inputLabelsImage <input label map image> \
 --inputLandmarksFile <input landmarks file (output of BCD)> \
 --outputTalairachLandmarksFile <output landmarks file (including AC,PC,SLA,IRP)>