BUG: Fix a bug in nrrd header change by gtractResampleDWIInPlace
When resampling the input DWI image, the gtractResampleDWIInPlace should
also change the content of gradient diffusions table by both the
measurement frame and the input transform file. However, previously the
gradient directions were not affected by the input transform.

Before fixing the above bug, I designed a test to catch that in future.
First, I added a new option to the gtractResampleDWIInPlace commandline.
If --writeOutputMetaData ./metaData.csv is specified in the command
line, then it writes out the output image measurement frame and diffusion
gradient directions in a simple CSV file. This helps to be able to verify the
output image meta data in TestSuite.

Then, I added new python program to compare two csv files (baseline and
input) and added a new test based on this new program.