ENH: Write output transforms of BFit and BABCD in float
By default the output registration transforms (either the output
composite transform or each transform component) are written to the
disk in double precision.
This patch provides a new flag to BRAINSFit called
"writeOutputTransformInFloat". If this flag is ON, the output transforms
will be written in single (float) precision. It is especially important
if the output transform is a displacement field transform, or it is a
composite transform that includes several displacement fields.