BUG: Fix dtiestim cmdline parameters
By default 'dtiestim' runs an OTSU threshold on the baseline image (B0
image), and this thresholding mask defines the voxels where the tensors
are computed.
Unfortunately this thresholding mask cause many brain voxels with
relatively low values are left out in the output tensor image and RISs
created from that.
Since we have a well defined mask for the brain region, such an
additional threshold masking is NOT desired, so we need to explicity use
the option --threshold 0.