ENH: Fixed automatic versioning
It is desirable to have automated versioning for BRAINSTools.
In order to have this, it was necessary to have code added
that figures out the nearest version to the current release.

In order for the cmake scripts to work, it was necessary to
have at least 1 version number taged in the repository.