BUG: Fixed DWIConverter based on last changes to itkDCMTKFileReader
Prevented GetOrigin bug originated from itkDCMTKFileReader

There is a bug in following function in ITK:
int DCMTKFileReader::GetOrigin(double *origin)

so after the first call of GetOrigin, origin is reset to zero.
This bug should be addressed in ITK; however, before that I prevented
this issue in DWIConverter.

ITK as of 2014-04-20 fixed the code in ITK, and that will prevent
the problem from occuring.