ENH: Fixed, enhanced and tested DWIMetaDataDictionaryValidator
This patch fixes all bugs in DWIMetaDataDictionaryValidator.
Note that the following fields do not need to be set by Metadata
validator, since itkNrrdImageIO extract them from image information:
- type e.g. short
- dimension i.e. 4
- space i.e. left-posterior-superior
- sizes
- space directions
- kinds
- endian
- encoding
- space origin

Also, "space units" is an optional field, and ITK does not set that
in the output MetaDataDictionary even if it exists in the header of
read Nrrd file. Therefore, this field does not need to be set by
the validator.

The following fields should be managed by validator:
- thickness
- centerings
- measurement frame
- modality
- b-value
- gradient_xxxx

However, NOTE: "centerings" and "thickness" should be created based on
"volume interleaved".  If the image is "pixel interleave" (like vectorImage in ITK),
the NrrdIO will automatically handle the correct permutation.

------
Finally, DWIMetaDataDictionaryValidatorTest was changed to an ITK test
format and the input reference volume is compared with replicated
reference volume that is written with a new MetaDataDictionary.