BUG: Fixed FSLToNrrd converter
This patch uses the newly developed DWIMetaDataDictionaryValidator to
fix writing header of output nrrd file in FSLToNrrd converter.

The input FSL Nifti image is read as a "volume interleaved" image
(i.e. itk::Image<PixelType, 4>). Then, it is converted to a "pixel
interleaved" image by converting the ITK 4D volume to an
itkVectorImageType (i.e. itk::VectorImage<PixelType, 3>).

Finally the MetaDataDictionary created from
DWIMetaDataDictionaryValidator is added to the output VectorImage, and
it is written to the disk in Nrrd format.