ENH: Provided build consistency with c++03
DWIMetaDataDictionaryValidator class was using std::array that is only
available in c++11. However, we needed to be able to build with c++03,
since Slicer does not compile with c++11.
Instead of using std::array, I added a wrapper class for std::vector
that makes sure that vectors are defined with fixed size in DWI validator
class.