BUG: BCD could not process float image type
The image reader in BCD was templated based on 'short' voxel type.
It was causing incorrect results when the input image has 'float' voxel
type. To fix the issue, all input images are read by a float reader;
then their values are rescaled to 'short' range and the read image is
casted to 'short' type for further processing by BCD.