ENH: Generate pure plugs mask and its incorporation in BABC
This patch applies following changes to BRAINSABC:

1) Bias field correction and classification processes (EM and KNN)
are now run in physical space rather than voxel space.

2) A new feature is added to BRAINSABC that helps the program find the
voxels composed of partial volume effect, so only pure samples can be
used for classification and bias field correction.

An ITK membership filter is added to identify pure samples. This
filter evaluates a multimodal sample set at each plug area to check
whether all the picked samples within a plug area are affected by partial
volume issue or not based on a user defined threshold.

This threshold is a value between zero and one specified by a newly added
flag named "--purePlugsThreshold". The default value is zero that means
not using pure plugs for classification.
If you want to activate using pure plugs for EM computations and KNN
training, a value of 0.2 is suggested empirically.
The threshold value defines a Mahalanobis-weighted Euclidean distance
that is used as the metric in a ITK membership filter.

Size of a plug area is specified automatically as the lowest possible
resolution based on the voxel size of input multimodal scans.

The number of samples at each plug area is defined by a new BRAINSABC
flag named "--numberOfSubSamplesInEachPlugArea". The default value to
this flag is 0,0,0 that means number of picked samples will be decided
based on the ratio of finest to lowest resolution at each coordinate
direction.

Technically if you wish to run BRAINSABC using pure plugs feature, set a
value greater than zero (0.2 has shown promising results) to
"purePlugsThreshold".

----
If you want to generate a pure plug mask outside of BRAINSABC, use the
newly added program called "GeneratePurePlugMask".
Note that it is assumed that input multimodal scans to this program are
already aligned in physical space.

Usage:
$ bin/GeneratePurePlugMask \
  --inputImageModalities t1.nii.gz,t2.nii.gz,IDWI.nrrd \
  --threshold 0.2 \
  --numberOfSubSamples 2,2,2 \
  --outputMaskFile binary_pure_mask.nrrd

GeneratePurePlugMask.h is called to compute pure plugs, and this file
calls "itkIntegrityMetricMembershipFunction" filter to decide each plug
is pure or not.