COMP: Fixes some CMake dependency issues with the FindVTKUtil module
If a specific module was not listed in Common.cmake to turn on the
BRAINSTools_REQUIRES_VTK variable, then calling the FindVTKUtil macro
would do nothing, and no errors or warnings would be generated. Now if
a module tries to use FindVTKUtil but it isn't on the list, then there
will be an error with an explanation. Also, two modules were added to
the list to turn on BRAINSTools_REQUIRES_VTK