ENH: run multiresolution AtlasToSubject SyN registration
In BRAINSABC, when AtlasToSubject transform choice is set to SyN,
and the SyN transform is restored from an already saved state, we need to only
update the state by running SyN registration only in full resolution.
However, if SyN stage is initialized NOT from the state, we need to run
SyN registration in mulitple resolution levels.