BUG: Fix failing BCD tests by adjusting image compare tolerences
BCD tests were passing on mac but failing on linux. Allowing more
differences between the baseline and test images made these tests
pass. A typicall range and size of intensity values for the baselines
is 1500 and 256*256*256, so setting compareIntensityTolerance to 50
and compareNumberOfPixelsTolerence to 12 allows a small and reasonable
variation between different machines and platforms.