BUG: BRAINSResample: Fix tests removing usage of dynamic_cast with transform
This commit fixes a regression introduced
by InsightSoftwareConsortium/ITK@d1e9fe2 (BUG: Ensure Transform IO
factories are  registered once.)

Following this ITK commit, the ITKIO transform libraries are build as
shared and dynamic_cast can NOT be used with templated classes that are
instantiated in a translation unit different than the one where they are
defined. It will work only if the classes are explicitly instantiated
and exported.

See InsightSoftwareConsortium/ITK@d1e9fe2
and see http://stackoverflow.com/questions/8024010/why-do-template-class-functions-have-to-be-declared-in-the-same-translation-unit

This commit fixes the following tests compiled using gcc 4.4.3
and executed on Ubuntu 10.04:

  ValidateBRAINSResampleTest4_nii
  ValidateBRAINSResampleTest7_nii
  ValidateBRAINSResampleTest6_nii

Runtime error reported was the following:

631: ******* HERE *******/path/to/Slicer-0-build/BRAINSTools/BRAINSResample/BRAINSResample.cxx 197
631: itk::ExceptionObject (0x26ca9c0)
631: Location: "unknown"
631: File: /path/to/Slicer-0-build/BRAINSTools/BRAINSCommonLib/GenericTransformImage.cxx
631: Line: 385
631: Description: itk::ERROR: Error in type conversion