ENH: Add grainsize explicitly for blocked_ranges
By default the grainsize for blocked_ranges is
set to 1.  In many cases this is a reasonable default,
but in 3d ranges, it often makes sense for the grainsize
to be 1 for the outter loop, and less for inner loops.

As a heuristic, the outter loops (pages) have a grainsize
set to 1, the first inner loop (rows) is set to size[1]/2,
and the most inner loop (cols) is set to 512 (so there is no
splitting unles the image is very large).