BUG: defined an orderedmap class for BABC
When BABC is run using multimodal scans, it is important that they are
saved in a map based on a proper order i.e. T1 type always comes first,
then T2, PD, Flair, and Others.
The current implementation was using a std::map container for holding
the input types, so they were ordered alphabetically and PD was
considered as the first (main) input modality.