BUG: fix centerOfROI initializer
Implementation of centerOfROI initializer was not done migrated correctly from
the ITKv3 version of BRAINSFit. The initializer is supposed to work by first
aligning the centers of the user-defined masks, and then performing the sparse
search, the same way as centerOfHead initializer works. This is how it was implemented
in the pre-migration code, see:

https://github.com/fedorov/BRAINSFitIGT/blob/master/BRAINSCommonLib/BRAINSFitHelperTemplate.txx#L163

Instead, the erroneous implementation did just the alignment of the mask centers
using masks' moments.

This commit should fix the problem.