ENH: Added GenerateEdgeMapImage for super resolution
A new directory called "BRAINSSuperResolution" is added to the
BRAINSTools. This directory can be the home of an under-developing tool
for super resolution reconstruction of 3D volumes that is applicable to
4D DWI images.

First step of super resolution reconstruction through a weighted TV
approach is creating an edgemap from available high resolution
modalities.

A new program called "GenerateEdgeMapImage" is added.
Usage:

GenerateEdgeMapImage  \
--inputMRVolumes {list of input MR volumes in higher spatial resolution}
--lowerPercentileMatching 0.50 \
--upperPercentileMatching 0.95
--maximumOutputRange 255 \
--minimumOutputRange 1 \
--inputMask {If set, image histogram percentiles will be calculated
  within the mask} \
--outputEdgeMap EdgeMap.nrrd \
--outputMaximumGradientImage MaximumGradientImage.nrrd