ENH: Throw IO Exception when reading/writing invalid landmarks file
The functions provided by Slicer3LandmarkIO allow landmark files to be written
and read.
Previously, if an invalid filename for a landmarks file was supplied, an error
message was printed, and if reading, an empty set of landmarks was returned.
This could cause problems, as it would be difficult to know if the landmarks
file was empty, or if an invalid filename was supplied. By throwing an exception
a program calling the function can halt progress if necessary. This will also
make debugging much easier.