BUG: Converting floating point data can cause overflow problems
The output data type is short int for DWI data, so if the input data is of type float, overflow
can occur that cause invalid data results to be generated.  This patch set identifies
input data types as float, and then rescales the intensities to positive short dynamic range.

This resolves #317.