BUG:  Failure tolerance was too narrow
In a very few test cases, the failure tolerance for
where to find the eyes of a subject was too small.

This tolerance has been slightly increased to allow
for identifying eyes in the few subjects that were
failing to run due to slightly unexpected eye locations.

When an invalid condition is identified, provide
starting files for manually fixing and re-running
data analysis.