BUG: Nipype needs OrderedDict for hash consistency
When using a dictionary in a nipype wrapped python
code or as inputs/outputs to nodes, it is important
to use OrderedDict so that consistent command line
hashing is maintained.

OrderedDicts do not pose a performance concern
for our uses, so globally replace all dict()
with OrderedDict() as a safety precaution.