COMP: Add quick -march=native compiler flag adding
In many cases building with -march=native can drastically increase
the runtime performance of algorithms.  When it is known
that a binary set will only run on one platform, then
one can use cmake -DBUILD_OPTIMIZED:BOOL=ON to quickly
set the compiler flags for all projects to use the same
optimized settings.