ENH: Add references to moving and fixed volumes from the output transform
Sometimes it is important to be able to identify the sources of an output, for example in this case the input moving and fixed volumes from the transform, or the original moving volume from the output volume (i.e. resampled moving volume).

A new CLI feature makes this possible, see https://github.com/Slicer/SlicerExecutionModel/pull/109 and https://github.com/Slicer/Slicer/pull/1120. This commit adds these references to BRAINSFit.