STYLE: Assign initial values to variables
Checks whether there are local variables that are declared without an initial
value. These may lead to unexpected behaviour if there is a code path that
reads the variable before assigning to it.

Only integers, booleans, floats, doubles and pointers are checked. The fix
option initializes all detected values with the value of zero. An exception is
float and double types, which are initialized to NaN.

cd
run-clang-tidy.py -extra-arg=-D__clang__ -checks=-*,cppcoreguidelines-init-variables  -header-filter=.* -fix
# https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-init-variables.html